+++ client/src/components/FilterList/index.jsx
@@ -1,0 +1,11 @@
+function FilterList({ filterData }) {
+  return filterData.length
+    ? filterData.map((i, j) => (
+        <li key={j}>
+          {i.filterName} ({i.filterType})
+        </li>
+      ))
+    : '';
+}
+
+export default FilterList;