+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 33. Render Locations on Profile page using React component
 34. Render Instruments on Profile Page using React component
 35. Update NavBar component to use DropDown sp Profile Tabs can be split into separate Menu Options
+36. Render Filters on Profile Page using React component